+++ src/pages/options/index.tsx
@@ -3,12 +3,12 @@ import React, { useCallback, useEffect, useRef, useState } from 'react';
 import { css } from '@emotion/css';
 import { IconFolderOpen, IconHelpCircle, IconMenu, IconSetting } from '@douyinfe/semi-icons';
 import { useGetState, useResponsive } from 'ahooks';
-import { convertToRule } from '@/share/core/ruleUtils';
+import { convertToRule } from '@/share/core/rule-utils';
 import { t } from '@/share/core/utils';
-import { prefs } from '@/share/core/storage';
-import { Rule } from '@/share/core/var';
+import { prefs } from '@/share/core/prefs';
+import type { Rule } from '@/share/core/types';
 import SemiLocale from '@/share/components/semi-locale';
-import isDarkMode from '@/share/components/is-dark-mode';
+import isDarkMode from '@/share/pages/is-dark-mode';
 import GroupSelect from './sections/groupSelect';
 import ImportAndExportSection from './sections/importAndExport';
 import OptionsSection from './sections/options';
