+++ src/pages/options/sections/rules/edit/index.tsx
@@ -1,15 +1,15 @@
-import { BoolRadioGroupField } from '@/pages/options/components/bool-radio';
-import Api from '@/share/core/api';
-import emitter from '@/share/core/emitter';
-import { initRule, isMatchUrl } from '@/share/core/ruleUtils';
-import { prefs } from '@/share/core/storage';
-import { IS_SUPPORT_STREAM_FILTER, t } from '@/share/core/utils';
-import { InitdRule, IS_MATCH, Rule, RULE_MATCH_TYPE, RULE_TYPE } from '@/share/core/var';
 import { IconSave } from '@douyinfe/semi-icons';
 import { Button, Form, Input, SideSheet, Toast } from '@douyinfe/semi-ui';
 import { FormApi } from '@douyinfe/semi-ui/lib/es/form';
 import { css } from '@emotion/css';
 import * as React from 'react';
+import { BoolRadioGroupField } from '@/pages/options/components/bool-radio';
+import Api from '@/share/pages/api';
+import { initRule, isMatchUrl } from '@/share/core/rule-utils';
+import { prefs } from '@/share/core/prefs';
+import { IS_SUPPORT_STREAM_FILTER, t } from '@/share/core/utils';
+import type { InitdRule, Rule } from '@/share/core/types';
+import { IS_MATCH, RULE_MATCH_TYPE, RULE_TYPE } from '@/share/core/constant';
 import { CodeEditorField } from './code-editor';
 import ENCODING_LIST from './encoding';
 import COMMON_HEADERS from './headers';
