+++ src/pages/options/sections/rules/utils.ts
@@ -1,8 +1,9 @@
-import Api from '@/share/core/api';
-import file from '@/share/core/file';
-import { createExport } from '@/share/core/ruleUtils';
+import Api from '@/share/pages/api';
+import { TABLE_NAMES_ARR } from '@/share/core/constant';
+import file from '@/share/pages/file';
+import { createExport } from '@/share/core/rule-utils';
 import { getTableName } from '@/share/core/utils';
-import { Rule, TABLE_NAMES } from '@/share/core/var';
+import type { Rule } from '@/share/core/types';
 import { getExportName } from '../../utils';
 
 export function toggleRule(rule: Rule, enable: boolean) {
@@ -21,7 +22,7 @@ export function save(rule: Rule) {
 
 export function batchShare(rules: Rule[]) {
   const result: any = {};
-  TABLE_NAMES.forEach((tb) => {
+  TABLE_NAMES_ARR.forEach((tb) => {
     result[tb] = [];
   });
   rules.forEach((e) => result[getTableName(e.ruleType)].push(e));
