+++ tsconfig.json
@@ -5,11 +5,11 @@
     "baseUrl": ".",
     "outDir": "build",
     "module": "esnext",
-    "target": "ES2017",
+    "target": "es2020",
     "jsx": "react",
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
-    "lib": ["ES2017", "dom"],
+    "lib": ["es2020", "dom"],
     "sourceMap": true,
     "allowJs": true,
     "rootDir": "./",
